pyproject.toml
[project]
name = "speacher"
version = "0.1.0"
description = "Multi-cloud audio transcription service with speaker diarization"
authors = [
    {name = "Rafal Lagowski", email = "rafal.lagowski@accenture.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core dependencies
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "websockets==12.0",
    "python-multipart==0.0.6",
    "pymongo==4.6.0",
    "pydantic==2.5.0",
    "httpx==0.25.1",
    # Frontend
    "streamlit==1.28.2",
    "streamlit-webrtc==0.47.1",
    "pandas==2.1.3",
    "requests==2.31.0",
    "numpy>=1.26.0,<2.0",
    "av>=13.0.0",
    # Cloud services - AWS
    "boto3==1.33.7",
    "botocore==1.33.7",
    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0,<3.0.0",
    # Cloud services - Azure
    "azure-storage-blob==12.19.0",
    "azure-cognitiveservices-speech==1.34.0",
    # Cloud services - Google Cloud
    "google-cloud-speech==2.22.0",
    "google-cloud-storage==2.13.0",
    # Utilities
    "python-dotenv==1.0.0",
    "typing-extensions==4.8.0",
    "pyjwt>=2.10.1",
    "passlib>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "email-validator>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "pytest-asyncio==0.21.1",
    "black==23.11.0",
    "flake8==6.1.0",
    "mypy==1.7.0",
    "isort==5.12.0",
    "mongomock==4.1.2",
]

test = [
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "pytest-asyncio==0.21.1",
    "mongomock==4.1.2",
]

[project.scripts]
speacher = "speacher.cli:main"
speacher-api = "backend.main:run_server"
speacher-ui = "frontend.app:run_app"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
namespace_packages = false
explicit_package_bases = true
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=src --cov-report=term-missing"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/speacher", "src/backend"]

[dependency-groups]
dev = [
    "mongomock>=4.1.2",
    "setuptools>=80.9.0",
]
